hostelhub/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.1
email_validator==2.1.1
python-dateutil==2.9.0
Flask-Login==0.6.3

